Add a mechanism to print the path to all instances of a certain type
This is very prototype but it lets you see quickly where memory is being
held. Ideally this could be expanded to create a graph (dot viz?)
showing who is holding memory for all types.

Really this would be best if we didn't have to add custom code to scan a
specific field, better if we could just scan everything from the root.

Maybe this could be an addition to go tool pprof?